ml/requirements.txt
scikit-learn==1.2.2
catboost==1.2.8
joblib==1.5.1
numpy==1.26.4
pandas==2.2.3
scipy==1.15.3
websocket-client
rel